+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # RealmJoin Runbooks Changelog 
 
+## 2023-05-11
+* Convert to shared mailbox: Skip removing on-prem synced groups
+
 ## 2023-04-25
 * New Runbook: Create an Application Registration